Adjust maxDepth param to be float

+++ src/Macros/Recursive.php
@@ -7,9 +7,9 @@ use Illuminate\Support\Collection;
 /**
  * Recursivly convert arrays and objects within a multi-dimensional array to Collections
  *
- * @param ?int $maxDepth
+ * @param ?float $maxDepth
  * @param ?int $depth
- * 
+ *
  * @mixin \Illuminate\Support\Collection
  *
  * @return \Illuminate\Support\Collection
@@ -18,7 +18,7 @@ class Recursive
 {
     public function __invoke()
     {
-        return function (?int $maxDepth = INF, ?int $depth = 0): Collection {
+        return function (?float $maxDepth = INF, ?int $depth = 0): Collection {
             return $this->map(function ($value) use ($depth, $maxDepth) {
                 if ($depth > $maxDepth) {
                     return $value;
